requirements.txt
# ============================================================
# JARVIS - Requirements.txt COMPLET ET FIXÉ
# ============================================================
# Installation : pip install -r requirements.txt

# ============================================================
# LLM Local
# ============================================================
ollama>=0.1.6

# ============================================================
# Speech-to-Text (STT)
# ============================================================
faster-whisper>=0.10.0
openai-whisper>=20231117
webrtcvad>=2.0.10

# ============================================================
# Transformers - VERSION FIXE CRITIQUE
# ============================================================
# NE PAS MODIFIER - Requis pour BeamSearchScorer
transformers==4.33.0

# ============================================================
# Text-to-Speech (TTS)
# ============================================================
edge-tts>=6.1.9  # Microsoft TTS
pygame>=2.5.2     # Pour lecture audio Edge-TTS

# ============================================================
# Voice Cloning - VERSIONS CRITIQUES
# ============================================================
# PyTorch - VERSION FIXE pour éviter warnings weights_only
torch==2.1.0
torchaudio==2.1.0
torchcodec==0.8

# Coqui TTS pour clonage vocal
TTS>=0.22.0

# ============================================================
# Audio Processing
# ============================================================
pyaudio>=0.2.13
sounddevice>=0.4.6
soundfile>=0.12.1
scipy>=1.10.0,<1.14.0
librosa==0.11.0
numphy==1.22.3
numba>=0.57.0

# ============================================================
# Web Framework
# ============================================================
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# ============================================================
# Utilities
# ============================================================
pyyaml>=6.0
colorama>=0.4.6
keyboard>=0.13.5
psutil>=5.9.6
Pillow>=10.0.0
num2words

# ============================================================
# Dépendances additionnelles TTS
# ============================================================
einops>=0.7.0
encodec>=0.1.1
inflect>=7.0.0
anyascii>=0.3.2
pydub>=0.25.1

# ============================================================
# NOTES IMPORTANTES
# ============================================================

# VERSIONS CRITIQUES À NE PAS MODIFIER :
# - transformers==4.33.0  → Pour BeamSearchScorer
# - torch==2.3.1          → Évite problème weights_only de PyTorch 2.6+
# - numpy<2.0.0           → TTS pas compatible numpy 2.0

# FFmpeg REQUIS :
# Windows : winget install "FFmpeg (Essentials Build)"
# Linux   : sudo apt install ffmpeg

# INSTALLATION COMPLÈTE :
# 1. python -m venv venv
# 2. venv\Scripts\activate
# 3. pip install -r requirements.txt
# 4. Installer FFmpeg (voir ci-dessus)